config: Update webpack for CSS and JS
Removed babel from JS loaders (done by local environment). On
production we also will do a babel build before running webpack
so no need to have it as part of the js loaders.

Added SASS loading for CSS and the extract text plugin for creating
the style bundle to use on the page. For now just having one bundle
but could possibly break it up for multiple pages.

see: #1